DiffChassis: fix wrong assignment in constructor and change all float to double.

+++ PrestavbaMob/Kod/DifferencialniPodvozek/mobDifferencialChassis.cpp
@@ -20,7 +20,6 @@ const unsigned int BUFFER_SIZE = 10;
 
 MobDifferentialChassis::MobDifferentialChassis(DiffChassisParam &diffChassisParam) : BasicDifferentialChassis(diffChassisParam) {
 	end_ = false;
-
 	sendMotorPower(motorsPower(0, 0));
 
 	loopPidThread_ = std::move(std::thread(&MobDifferentialChassis::updateEncoders,this,5));
@@ -40,11 +39,10 @@ DistanceWheels MobDifferentialChassis::computeDistance(Encoders distance) {
 	DistanceWheels distanceInMeters;
 	distanceInMeters.left = distance.left * diffChassisParam_.metersPerTick;
 	distanceInMeters.right = distance.right * diffChassisParam_.metersPerTick;
-
 	return distanceInMeters;
 }
 
-VelocityWheels MobDifferentialChassis::computeSpeed(DistanceWheels distance, float time) {
+VelocityWheels MobDifferentialChassis::computeSpeed(DistanceWheels distance, double time) {
 	VelocityWheels speed;
 	speed.left = distance.left / time;
 	speed.right = distance.right / time;
@@ -54,7 +52,7 @@ VelocityWheels MobDifferentialChassis::computeSpeed(DistanceWheels distance, flo
 
 void MobDifferentialChassis::changeRobotState(DistanceWheels change) {
 	double angleChange = (change.right - change.left) / diffChassisParam_.wheelbase;
-	double distanceChange = (change.right + change.left) / 2;
+	double distanceChange = (change.right + change.left) / 2.0;
 
 	std::lock_guard<std::mutex> lock(stateMutex_);
 	wheelsDistance_.left += change.left;
@@ -73,8 +71,8 @@ int MobDifferentialChassis::sendMotorPower(struct motorsPower speedMotors) {
 	return diffChassisParam_.driver->setMotorsPower(speedMotors);
 }
 
-int MobDifferentialChassis::PIRegulator(const float actualSpeed,const float desireSpeed, PIValue &PIParam) {
-	float speedDifference = desireSpeed - actualSpeed;
+int MobDifferentialChassis::PIRegulator(const double actualSpeed,const double desireSpeed, PIValue &PIParam) {
+	double speedDifference = desireSpeed - actualSpeed;
 	PIParam.ISum += speedDifference;
 
 	int speed = PIParam.P * speedDifference + PIParam.I * PIParam.ISum;
@@ -96,7 +94,7 @@ void MobDifferentialChassis::updateEncoders(const int period) {
 		Encoders differenceEncoders = getChangeOfEncoders();
 
 		// Compute actual speed and use PID
-		float time = (microStart - lastMicroTime) / 1000000.0f; // time in sec
+		double time = (microStart - lastMicroTime) / 1000000.0f; // time in sec
 		DistanceWheels distance = computeDistance(differenceEncoders);
 		VelocityWheels actualSpeed = computeSpeed(distance, time);
 		changeRobotState(distance);
@@ -132,15 +130,15 @@ void MobDifferentialChassis::stop(bool slow) {
 		do{
 			VelocityWheels desire;
 			if(wheelsSpeed_.left > 0){
-				desire.left = rob_fce::valueInRange<float>(wheelsSpeed_.left - STOP_SPEED, 0.0 , wheelsSpeed_.left);
+				desire.left = rob_fce::valueInRange<double>(wheelsSpeed_.left - STOP_SPEED, 0.0 , wheelsSpeed_.left);
 			}else{
-				desire.left = rob_fce::valueInRange<float>(wheelsSpeed_.left + STOP_SPEED, wheelsSpeed_.left, 0.0);
+				desire.left = rob_fce::valueInRange<double>(wheelsSpeed_.left + STOP_SPEED, wheelsSpeed_.left, 0.0);
 			}
 
 			if(wheelsSpeed_.right > 0){
-				desire.right = rob_fce::valueInRange<float>(wheelsSpeed_.right - STOP_SPEED, 0.0 , wheelsSpeed_.right);
+				desire.right = rob_fce::valueInRange<double>(wheelsSpeed_.right - STOP_SPEED, 0.0 , wheelsSpeed_.right);
 			}else{
-				desire.right = rob_fce::valueInRange<float>(wheelsSpeed_.right + STOP_SPEED, wheelsSpeed_.right, 0.0);
+				desire.right = rob_fce::valueInRange<double>(wheelsSpeed_.right + STOP_SPEED, wheelsSpeed_.right, 0.0);
 			}
 
 			setVelocity(desire);
@@ -164,14 +162,14 @@ void MobDifferentialChassis::setVelocity(const VelocityWheels speed) {
 	wheelsDesireSpeed_ = safeSpeed;
 }
 
-void MobDifferentialChassis::setVelocity(const float distance, const float angle){
+void MobDifferentialChassis::setVelocity(const double distance, const double angle){
 	VelocityWheels velocity;
 	if(std::abs(angle) < EPSILON){
 		velocity.left = diffChassisParam_.maxVelocity;
 		velocity.right = diffChassisParam_.maxVelocity;
 	}else{
-		float radius = std::abs(distance / angle);
-		float ratio = (radius - diffChassisParam_.wheelbase / 2.0) / (radius + diffChassisParam_.wheelbase / 2.0);
+		double radius = std::abs(distance / angle);
+		double ratio = (radius - diffChassisParam_.wheelbase / 2.0) / (radius + diffChassisParam_.wheelbase / 2.0);
 		if(angle >= 0){
 			velocity.left = diffChassisParam_.maxVelocity*ratio;
 			velocity.right = diffChassisParam_.maxVelocity;
